Add support for Ethernet LLM model now is downloaded to the FPGA board by Ethernet that comes with Arty devkit, extra SSD shield no longer required to run LLM example. LLM sampling algo is improved for better token generation.

+++ SW/apps/llm/kernels/llm_m.h
@@ -66,7 +66,7 @@ extern void kernel_llm_scale_exe(int reqId,int N,float16_t *x,float scale);
 
 extern int kernel_llm_find_max(float16_t *x,uint32_t N);
 
-extern int kernel_llm_find_k_max(float16_t *x,uint32_t N,int K,int *top,float *topp);
+extern int kernel_llm_find_k_max(float16_t *x,uint32_t N,int K,int *top,float16_t *topp);
 
 extern void kernel_llm_done();
